# [async] Add bit mask in the SFG node edges container