--- a/taichi/program/async_engine.cpp
+++ b/taichi/program/async_engine.cpp
@@ -208,7 +208,9 @@ void AsyncEngine::launch(Kernel *kernel, Context &context) {
 void AsyncEngine::synchronize() {
   TI_AUTO_PROF;
   bool modified = true;
-  sfg->rebuild_graph(/*sort=*/false);
+  sfg->reid_nodes();
+  sfg->reid_pending_nodes();
+  sfg->sort_node_edges();
   TI_TRACE("Synchronizing SFG of {} nodes ({} pending)", sfg->size(),
            sfg->num_pending_tasks());
   debug_sfg("initial");