@yuanming-hu 2020-12-07
The performance benchmark in #2090 shows very promising results:

Before this PR:
 - `total time 764.4606 ms; per_node 763.6969 ns`

After this PR (1.21x faster!):
 - `total time 631.5184 ms; per_node 630.8875 ns`